HDR/requirements.txt
flask
opencv-python-headless
numpy
Pillow